refactor: single shortcut handler for editor and slideshow
